Extract ICS/CalDAV parser from App.jsx to src/utils/icsParser.js
Moves parseICS (RRULE expansion, EXDATE handling, VEVENT RECURRENCE-ID
overrides, VTODO handling), parseDatetime, filterByDateWindow, and
expandMultiDayEvent out of App.jsx verbatim. All four are pure data
transformations with no React dependency — the only import is
dateToString from taskUtils. App.jsx shrinks by ~430 lines.

Adds a test suite covering line unfolding, text unescaping, VTODO
DUE handling, weekly RRULE expansion with EXDATE, recurrence override
move/cancel, the masterUids marker, UTC datetime conversion, retention
filtering, and multi-day event splitting.

Part of the App.jsx decomposition tracked in CLAUDE.md.
